+++ src/main/java/com/github/alexdlaird/ngrok/process/NgrokProcess.java
@@ -19,7 +19,10 @@ import com.github.alexdlaird.ngrok.protocol.Tunnels;
 
 import java.io.BufferedReader;
 import java.io.IOException;
+import java.io.InputStream;
 import java.io.InputStreamReader;
+import java.nio.charset.Charset;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.util.ArrayList;
 import java.util.Calendar;
@@ -37,8 +40,8 @@ import static java.util.Objects.nonNull;
 import static java.util.logging.Level.SEVERE;
 
 /**
- * An object containing information about the <code>ngrok</code> process.
- * Can be configured with {@link JavaNgrokConfig}.
+ * An object containing information about the <code>ngrok</code> process. Can be configured with
+ * {@link JavaNgrokConfig}.
  */
 public class NgrokProcess {
 
@@ -51,8 +54,8 @@ public class NgrokProcess {
 
     /**
      * If <code>ngrok</code> is not already installed at {@link JavaNgrokConfig#getNgrokPath()}, the given
-     * {@link NgrokInstaller} will install it. This will also provision a default <code>ngrok</code> config
-     * at {@link JavaNgrokConfig#getConfigPath()}, if none exists.
+     * {@link NgrokInstaller} will install it. This will also provision a default <code>ngrok</code> config at
+     * {@link JavaNgrokConfig#getConfigPath()}, if none exists.
      *
      * @param javaNgrokConfig The <code>java-ngrok</code> to use when interacting with the <code>ngrok</code> binary.
      * @param ngrokInstaller  The class used to download and install <code>ngrok</code>.
@@ -66,7 +69,8 @@ public class NgrokProcess {
             ngrokInstaller.installNgrok(javaNgrokConfig.getNgrokPath(), javaNgrokConfig.getNgrokVersion());
         }
         if (!Files.exists(javaNgrokConfig.getConfigPath())) {
-            ngrokInstaller.installDefaultConfig(javaNgrokConfig.getConfigPath(), Collections.emptyMap(), javaNgrokConfig.getNgrokVersion());
+            ngrokInstaller.installDefaultConfig(javaNgrokConfig.getConfigPath(), Collections.emptyMap(),
+                javaNgrokConfig.getNgrokVersion());
         }
     }
 
@@ -81,8 +85,9 @@ public class NgrokProcess {
         }
 
         if (!Files.exists(javaNgrokConfig.getNgrokPath())) {
-            throw new NgrokException(String.format("ngrok binary was not found. " +
-                    "Be sure to call \"NgrokInstaller.installNgrok()\" first for \"ngrokPath\": %s", javaNgrokConfig.getNgrokPath()));
+            throw new NgrokException(String.format("ngrok binary was not found. "
+                    + "Be sure to call \"NgrokInstaller.installNgrok()\" first for \"ngrokPath\": %s",
+                javaNgrokConfig.getNgrokPath()));
         }
         ngrokInstaller.validateConfig(javaNgrokConfig.getConfigPath());
 
@@ -112,7 +117,13 @@ public class NgrokProcess {
         processBuilder.command(command);
         try {
             process = processBuilder.start();
-            Runtime.getRuntime().addShutdownHook(new Thread(this::stop));
+            Runtime.getRuntime().addShutdownHook(new Thread(() -> {
+                try {
+                    stop();
+                } catch (final IOException ex) {
+                    LOGGER.log(Level.INFO, "An error occurred when stopping the process", ex);
+                }
+            }));
 
             LOGGER.fine("ngrok process starting");
 
@@ -138,9 +149,8 @@ public class NgrokProcess {
                 stop();
 
                 if (nonNull(processMonitor.startupError)) {
-                    throw new NgrokException(String.format("The ngrok process errored on start: %s.", processMonitor.startupError),
-                            processMonitor.logs,
-                            processMonitor.startupError);
+                    throw new NgrokException(String.format("The ngrok process errored on start: %s.",
+                        processMonitor.startupError), processMonitor.logs, processMonitor.startupError);
                 } else {
                     throw new NgrokException("The ngrok process was unable to start.", processMonitor.logs);
                 }
@@ -158,10 +168,10 @@ public class NgrokProcess {
     }
 
     /**
-     * Terminate the <code>ngrok</code> processes, if running. This method will not block, it will
-     * just issue a kill request.
+     * Terminate the <code>ngrok</code> processes, if running. This method will not block, it will just issue a kill
+     * request.
      */
-    public void stop() {
+    public void stop() throws IOException {
         if (!isRunning()) {
             LOGGER.info(String.format("\"ngrokPath\" %s is not running a process", javaNgrokConfig.getNgrokPath()));
 
@@ -172,18 +182,22 @@ public class NgrokProcess {
 
         processMonitor.stop();
         process.destroy();
+        if (nonNull(processMonitor.reader)) {
+            processMonitor.reader.close();
+        }
 
         process = null;
+        processMonitor = null;
     }
 
     /**
-     * Set the <code>ngrok</code> auth token in the config file, enabling authenticated features (for instance,
-     * more concurrent tunnels, custom subdomains, etc.).
+     * Set the <code>ngrok</code> auth token in the config file, enabling authenticated features (for instance, more
+     * concurrent tunnels, custom subdomains, etc.).
      *
      * <pre>
      * // Setting an auth token allows us to do things like open multiple tunnels at the same time
      * final NgrokClient ngrokClient = new NgrokClient.Builder().build();
-     * ngrokClient.setAuthToken("<NGROK_AUTHTOKEN>")
+     * ngrokClient.setAuthToken("&lt;NGROK_AUTHTOKEN&gt;")
      *
      * // &lt;NgrokTunnel: "http://&lt;public_sub1&gt;.ngrok.io" -&gt; "http://localhost:80"&gt;
      * final Tunnel ngrokTunnel1 = ngrokClient.connect();
@@ -193,8 +207,9 @@ public class NgrokProcess {
      *         .build();
      * final Tunnel ngrokTunnel2 = ngrokClient.connect(createTunnel);
      * </pre>
-     * <p>
-     * The auth token can also be set in the {@link JavaNgrokConfig} that is passed to the {@link NgrokClient.Builder}.
+     *
+     * <p>The auth token can also be set in the {@link JavaNgrokConfig} that is passed to the
+     * {@link NgrokClient.Builder}.
      *
      * @param authToken The auth token.
      */
@@ -226,9 +241,10 @@ public class NgrokProcess {
             final Process process = processBuilder.start();
             process.waitFor();
 
-            final BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream()));
-
+            final BufferedReader reader = new BufferedReader(
+                new InputStreamReader(process.getInputStream(), StandardCharsets.UTF_8));
             final String result = captureOutput(reader);
+            reader.close();
             if (!result.contains("Authtoken saved")) {
                 throw new NgrokException(String.format("An error occurred while setting the auth token: %s", result));
             }
@@ -241,7 +257,7 @@ public class NgrokProcess {
         final StringBuilder builder = new StringBuilder();
 
         String line;
-        while ((line = reader.readLine()) != null) {
+        while (nonNull(line = reader.readLine())) {
             builder.append(line).append("\n");
         }
 
@@ -288,9 +304,10 @@ public class NgrokProcess {
             final Process process = processBuilder.start();
             process.waitFor();
 
-            final BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream()));
-
+            final BufferedReader reader = new BufferedReader(
+                new InputStreamReader(process.getInputStream(), StandardCharsets.UTF_8));
             final String result = captureOutput(reader);
+            reader.close();
             return result.split("version ")[1];
         } catch (final IOException | InterruptedException | ArrayIndexOutOfBoundsException e) {
             throw new NgrokException("An error occurred while trying to update ngrok.", e);
@@ -323,6 +340,7 @@ public class NgrokProcess {
     }
 
     public static class ProcessMonitor implements Runnable {
+
         private final Process process;
         private final JavaNgrokConfig javaNgrokConfig;
         private final HttpClient httpClient;
@@ -330,6 +348,7 @@ public class NgrokProcess {
         private boolean tunnelStarted;
         private boolean clientConnected;
         private String startupError;
+        private BufferedReader reader;
 
         private final List<NgrokLog> logs = new ArrayList<>();
         private boolean alive = true;
@@ -350,10 +369,12 @@ public class NgrokProcess {
         @Override
         public void run() {
             try {
-                final BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream()));
+                final InputStreamReader inputStream = new InputStreamReader(process.getInputStream(),
+                    StandardCharsets.UTF_8);
+                reader = new BufferedReader(inputStream);
 
                 String line;
-                while ((line = reader.readLine()) != null) {
+                while (nonNull(line = reader.readLine())) {
                     logStartupLine(line);
 
                     if (isHealthy()) {
@@ -364,7 +385,9 @@ public class NgrokProcess {
                     }
                 }
 
-                while (alive && process.isAlive() && javaNgrokConfig.isKeepMonitoring() && (line = reader.readLine()) != null) {
+                while (alive && process.isAlive()
+                    && javaNgrokConfig.isKeepMonitoring()
+                    && nonNull(line = reader.readLine())) {
                     logLine(line);
                 }
 
@@ -403,7 +426,8 @@ public class NgrokProcess {
                 throw new JavaNgrokSecurityException(String.format("URL must start with \"http\": %s", apiUrl));
             }
 
-            final Response<Tunnels> tunnelsResponse = httpClient.get(String.format("%s/api/tunnels", apiUrl), Tunnels.class);
+            final Response<Tunnels> tunnelsResponse = httpClient.get(String.format("%s/api/tunnels", apiUrl),
+                Tunnels.class);
             if (tunnelsResponse.getStatusCode() != HTTP_OK) {
                 return false;
             }