+++ CHANGELOG.md
@@ -3,7 +3,12 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/alexdlaird/java-ngrok/compare/2.2.13...HEAD)
+## [Unreleased](https://github.com/alexdlaird/java-ngrok/compare/2.2.14...HEAD)
+
+## [2.2.14](https://github.com/alexdlaird/java-ngrok/compare/2.2.13...2.2.14) - 2024-03-06
+### Added
+- Build and stability improvements.
+- Documentation and style improvements.
 
 ## [2.2.13](https://github.com/alexdlaird/java-ngrok/compare/2.2.12...2.2.13) - 2024-02-26
 ### Added
@@ -163,8 +168,13 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Java 8 support.
 - `reconnectSessionRetries` from [`JavaNgrokConfig`](https://javadoc.io/doc/com.github.alexdlaird/java-ngrok/1.5.0/com/github/alexdlaird/ngrok/conf/JavaNgrokConfig.html), instead relying on `ngrok`'s own built-in retry mechanism on startup fails.
 
-## 1.4.x
+## [1.4.x](https://github.com/alexdlaird/java-ngrok/compare/1.4.12...1.4.x)
 The `1.4.x` branch is where support for Java 8 of `java-ngrok` is passively maintained. It is available through the `java8-ngrok` artifact on [Maven Central](https://maven-badges.herokuapp.com/maven-central/com.github.alexdlaird/java8-ngrok/).
+
+## [1.4.12](https://github.com/alexdlaird/java-ngrok/compare/1.4.11...1.4.12) - 2024-03-06
+### Added
+- Build and stability improvements.
+- Documentation and style improvements.
 
 ## [1.4.11](https://github.com/alexdlaird/java-ngrok/compare/1.4.8...1.4.11) - 2024-02-26
 ### Added